+++ Engine/src/function/impl/EnvironmentFunction.java
@@ -46,7 +46,7 @@ public class EnvironmentFunction extends AbstractFunction{
     }
 
     @Override
-    public Float getTicksNotUpdated(Context context) {
+    public Integer getTicksNotUpdated(Context context) {
         return null;
     }
