+++ Engine/src/function/impl/EvaluateFunction.java
@@ -39,7 +39,7 @@ public class EvaluateFunction extends AbstractFunction {
     }
 
     @Override
-    public Float getTicksNotUpdated(Context context) {
+    public Integer getTicksNotUpdated(Context context) {
         return null;
     }
 }
