+++ Engine/src/function/impl/RandomFunction.java
@@ -45,7 +45,7 @@ public class RandomFunction extends AbstractFunction {
     }
 
     @Override
-    public Float getTicksNotUpdated(Context context) {
+    public Integer getTicksNotUpdated(Context context) {
         return null;
     }
 }
